fix grepping of words that start with 'b'
try using <Leader>g to grep for a word that starts with 'b', e.g.
'boo', and you'll find that the regex that gets passed to grep is
actually "\boo\b" (should be "\bboo\b").

from `:h c_CTRL-R_CTRL-W`: "With CTRL-W the part of the word that was
already typed is not inserted again." -- vim thinks the 'b' in the
leading '\b' is the start of 'boo'. silly vim!